#165 [feat] : 캡슐화와 팩토리 패턴을 통한 SmallButton 디자인 시스템 구축, kdoc 주석
